# book-network-backend/src/main/java/dev/magadiflo/book/network/app/book/BorrowedBookResponse.java
package dev.magadiflo.book.network.app.book;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class BorrowedBookResponse {
    private Long id;
    private String title;
    private String authorName;
    private String isbn;
    private double rate;
    private boolean returned;
    private boolean returnApproved;
}