# book-network-backend/src/main/java/dev/magadiflo/book/network/app/feedback/FeedbackResponse.java
package dev.magadiflo.book.network.app.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {
    private Double note;
    private String comment;
    private boolean ownFeedback;
}